test: split file watcher coverage
